llm-evaluation/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
aiohttp>=3.8.0
scipy>=1.10.0
python-dotenv>=1.0.0
kaleido>=0.2.1  # For plotly image export
google-generativeai>=0.8.5
